+++ package.json
@@ -35,12 +35,12 @@
     "samples-test": "cd samples/ && npm link ../ && npm test && cd ../",
     "test-no-cover": "mocha build/test",
     "test": "npm run cover",
-    "lint": "eslint 'samples/*.js' 'samples/**/*.js'",
+    "lint": "eslint '**/*.js'",
     "docs": "jsdoc -c .jsdoc.js",
-    "fix": "eslint --fix 'samples/*.js' 'samples/**/*.js'",
+    "fix": "eslint --fix '**/*.js'",
     "generate-scaffolding": "repo-tools generate all && repo-tools generate lib_samples_readme -l samples/ --config ../.cloud-repo-tools.json",
     "clean": "gts clean",
-    "compile": "tsc -p . && cp -r src/v1/ build/src/v1/ && cp -r protos build/",
+    "compile": "tsc -p . && cp -r src/v1/ build/src/v1/ && cp -r protos build/ && cp smoke-test/*.js build/system-test && cp test/*.js build/test",
     "prepare": "npm run compile",
     "pretest": "npm run compile"
   },
