+++ samples/.eslintrc.yml
@@ -3,4 +3,4 @@ env:
   mocha: true
 rules:
   no-console: off
-  no-unused-vars: off
+  node/no-missing-require: off
