+++ smoke-test/.eslintrc.yml
@@ -2,5 +2,5 @@
 env:
   mocha: true
 rules:
-  node/no-unpublished-require: off
+  node/no-missing-require: off
   no-console: off
