+++ src/v1/publisher_client.js
@@ -109,9 +109,7 @@ class PublisherClient {
       topicPathTemplate: new gax.PathTemplate(
         'projects/{project}/topics/{topic}'
       ),
-      projectPathTemplate: new gax.PathTemplate(
-        'projects/{project}'
-      ),
+      projectPathTemplate: new gax.PathTemplate('projects/{project}'),
     };
 
     // Some of the methods on this service return "paged" results,
@@ -131,25 +129,36 @@ class PublisherClient {
     };
     let protoFilesRoot = new gax.GoogleProtoFilesRoot();
     protoFilesRoot = protobuf.loadSync(
-      path.join(__dirname, '..', '..', 'protos', 'google/iam/v1/iam_policy.proto'),
+      path.join(
+        __dirname,
+        '..',
+        '..',
+        'protos',
+        'google/iam/v1/iam_policy.proto'
+      ),
       protoFilesRoot
     );
     protoFilesRoot = protobuf.loadSync(
-      path.join(__dirname, '..', '..', 'protos', 'google/pubsub/v1/pubsub.proto'),
+      path.join(
+        __dirname,
+        '..',
+        '..',
+        'protos',
+        'google/pubsub/v1/pubsub.proto'
+      ),
       protoFilesRoot
     );
-
 
     // Some methods on this API support automatically batching
     // requests; denote this.
     this._descriptors.batching = {
       publish: new gax.BundleDescriptor(
         'messages',
-        [
-          'topic',
-        ],
+        ['topic'],
         'messageIds',
-        gax.createByteLengthFunction(protoFilesRoot.lookup('google.pubsub.v1.PubsubMessage'))
+        gax.createByteLengthFunction(
+          protoFilesRoot.lookup('google.pubsub.v1.PubsubMessage')
+        )
       ),
     };
 
@@ -190,7 +199,8 @@ class PublisherClient {
             }
         ),
         defaults[methodName],
-        this._descriptors.page[methodName] || this._descriptors.batching[methodName]
+        this._descriptors.page[methodName] ||
+          this._descriptors.batching[methodName]
       );
     }
 
@@ -222,7 +232,8 @@ class PublisherClient {
             }
         ),
         defaults[methodName],
-        this._descriptors.page[methodName] || this._descriptors.batching[methodName]
+        this._descriptors.page[methodName] ||
+          this._descriptors.batching[methodName]
       );
     }
   }
@@ -647,7 +658,7 @@ class PublisherClient {
       request,
       options
     );
-  };
+  }
 
   /**
    * Lists the names of the subscriptions on this topic.
@@ -741,7 +752,11 @@ class PublisherClient {
     }
     options = options || {};
 
-    return this._innerApiCalls.listTopicSubscriptions(request, options, callback);
+    return this._innerApiCalls.listTopicSubscriptions(
+      request,
+      options,
+      callback
+    );
   }
 
   /**
@@ -798,7 +813,7 @@ class PublisherClient {
       request,
       options
     );
-  };
+  }
 
   /**
    * Deletes the topic with the given name. Returns `NOT_FOUND` if the topic
@@ -1052,9 +1067,7 @@ class PublisherClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromTopicName(topicName) {
-    return this._pathTemplates.topicPathTemplate
-      .match(topicName)
-      .project;
+    return this._pathTemplates.topicPathTemplate.match(topicName).project;
   }
 
   /**
@@ -1065,9 +1078,7 @@ class PublisherClient {
    * @returns {String} - A string representing the topic.
    */
   matchTopicFromTopicName(topicName) {
-    return this._pathTemplates.topicPathTemplate
-      .match(topicName)
-      .topic;
+    return this._pathTemplates.topicPathTemplate.match(topicName).topic;
   }
 
   /**
@@ -1078,11 +1089,8 @@ class PublisherClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromProjectName(projectName) {
-    return this._pathTemplates.projectPathTemplate
-      .match(projectName)
-      .project;
+    return this._pathTemplates.projectPathTemplate.match(projectName).project;
   }
 }
-
 
 module.exports = PublisherClient;
