+++ src/v1/subscriber_client.js
@@ -112,9 +112,7 @@ class SubscriberClient {
       topicPathTemplate: new gax.PathTemplate(
         'projects/{project}/topics/{topic}'
       ),
-      projectPathTemplate: new gax.PathTemplate(
-        'projects/{project}'
-      ),
+      projectPathTemplate: new gax.PathTemplate('projects/{project}'),
       snapshotPathTemplate: new gax.PathTemplate(
         'projects/{project}/snapshots/{snapshot}'
       ),
@@ -179,7 +177,8 @@ class SubscriberClient {
             }
         ),
         defaults[methodName],
-        this._descriptors.page[methodName] || this._descriptors.stream[methodName]
+        this._descriptors.page[methodName] ||
+          this._descriptors.stream[methodName]
       );
     }
 
@@ -219,7 +218,8 @@ class SubscriberClient {
             }
         ),
         defaults[methodName],
-        this._descriptors.page[methodName] || this._descriptors.stream[methodName]
+        this._descriptors.page[methodName] ||
+          this._descriptors.stream[methodName]
       );
     }
 
@@ -664,7 +664,7 @@ class SubscriberClient {
       request,
       options
     );
-  };
+  }
 
   /**
    * Deletes an existing subscription. All messages retained in the subscription
@@ -1137,7 +1137,7 @@ class SubscriberClient {
       request,
       options
     );
-  };
+  }
 
   /**
    * Creates a snapshot from the requested subscription.<br><br>
@@ -1644,8 +1644,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromSubscriptionName(subscriptionName) {
-    return this._pathTemplates.subscriptionPathTemplate
-      .match(subscriptionName)
+    return this._pathTemplates.subscriptionPathTemplate.match(subscriptionName)
       .project;
   }
 
@@ -1657,8 +1656,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the subscription.
    */
   matchSubscriptionFromSubscriptionName(subscriptionName) {
-    return this._pathTemplates.subscriptionPathTemplate
-      .match(subscriptionName)
+    return this._pathTemplates.subscriptionPathTemplate.match(subscriptionName)
       .subscription;
   }
 
@@ -1670,9 +1668,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromTopicName(topicName) {
-    return this._pathTemplates.topicPathTemplate
-      .match(topicName)
-      .project;
+    return this._pathTemplates.topicPathTemplate.match(topicName).project;
   }
 
   /**
@@ -1683,9 +1679,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the topic.
    */
   matchTopicFromTopicName(topicName) {
-    return this._pathTemplates.topicPathTemplate
-      .match(topicName)
-      .topic;
+    return this._pathTemplates.topicPathTemplate.match(topicName).topic;
   }
 
   /**
@@ -1696,9 +1690,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromProjectName(projectName) {
-    return this._pathTemplates.projectPathTemplate
-      .match(projectName)
-      .project;
+    return this._pathTemplates.projectPathTemplate.match(projectName).project;
   }
 
   /**
@@ -1709,9 +1701,7 @@ class SubscriberClient {
    * @returns {String} - A string representing the project.
    */
   matchProjectFromSnapshotName(snapshotName) {
-    return this._pathTemplates.snapshotPathTemplate
-      .match(snapshotName)
-      .project;
+    return this._pathTemplates.snapshotPathTemplate.match(snapshotName).project;
   }
 
   /**
@@ -1722,11 +1712,9 @@ class SubscriberClient {
    * @returns {String} - A string representing the snapshot.
    */
   matchSnapshotFromSnapshotName(snapshotName) {
-    return this._pathTemplates.snapshotPathTemplate
-      .match(snapshotName)
+    return this._pathTemplates.snapshotPathTemplate.match(snapshotName)
       .snapshot;
   }
 }
-
 
 module.exports = SubscriberClient;
