--- system-test/.eslintrc.yml
@@ -1,6 +1,0 @@
----
-env:
-  mocha: true
-rules:
-  node/no-unpublished-require: off
-  no-console: off
