+++ test/.eslintrc.yml
@@ -2,4 +2,4 @@
 env:
   mocha: true
 rules:
-  node/no-unpublished-require: off
+  node/no-missing-require: off
